+++ march-randomness/round7.c
@@ -16,7 +16,7 @@ int main(void) {
   seed_rand();
 
   long times_same = 0;
-  long trials = 1e+8; // takes about 5 seconds to run
+  long trials = (long)1e+8; // takes about 5 seconds to run
 
   clock_t start = clock();
 
@@ -55,7 +55,7 @@ void seed_rand(void) {
   fread(&rand_seed, sizeof(rand_seed), sizeof(char), fp);
   fclose(fp);
   printf("random seed: %llu\n", rand_seed);
-  srand(rand_seed);
+  srand((unsigned int)rand_seed);
 }
 
 int int_rand(int min, int max) {
