+++ march-randomness/round9.c
@@ -21,7 +21,7 @@ int main(void) {
   seed_rand();
 
   long times_where_four = 0;
-  long trials = 1e+7; // takes about 2.5 seconds
+  long trials = (long)1e+7; // takes about 2.5 seconds
 
   clock_t start = clock();
 
@@ -62,7 +62,7 @@ void seed_rand(void) {
   fread(&rand_seed, sizeof(rand_seed), sizeof(char), fp);
   fclose(fp);
   printf("random seed: %llu\n", rand_seed);
-  srand(rand_seed);
+  srand((unsigned int)rand_seed);
 }
 
 int int_rand(int min, int max) {
